+++ kafka/net/__init__.py
@@ -3,6 +3,7 @@ from .inet import create_connection, KafkaNetSocket
 from .manager import KafkaConnectionManager
 from .metrics import KafkaConnectionMetrics, KafkaManagerMetrics
 from .selector import NetworkSelector
+from .http_connect import HttpConnectProxy
 from .socks5 import Socks5Proxy
 from .transport import KafkaTCPTransport, KafkaSSLTransport
 from .wakeup_notifier import WakeupNotifier
@@ -13,6 +14,6 @@ from .compat import KafkaNetClient
 __all__ = [
     'KafkaConnection', 'create_connection', 'KafkaNetSocket',
     'KafkaConnectionManager', 'KafkaConnectionMetrics', 'KafkaManagerMetrics',
-    'NetworkSelector', 'Socks5Proxy', 'KafkaTCPTransport', 'KafkaSSLTransport',
+    'NetworkSelector', 'HttpConnectProxy', 'Socks5Proxy', 'KafkaTCPTransport', 'KafkaSSLTransport',
     'WakeupNotifier', 'KafkaNetClient',
 ]